fix linking in telemetry-ffi
