+++ ddtelemetry/Cargo.toml
@@ -22,15 +22,14 @@ serde = {version = "1.0", features = ["derive"]}
 serde_json = {version = "1.0"}
 spawn_worker = {path = "../spawn_worker"}
 sys-info = {version = "0.9.0"}
-tarpc = {path = "/home/pawel/repos/tarpc/tarpc", features = ["serde-transport"]}
 tokio = {version = "1.23", features = ["sync", "io-util", "signal"]}
 tokio-serde = {version = "0.8", features = ["messagepack", "json"]}
-tokio-util = {version = "0.6.9", features = ["codec"]}
+tokio-util = {version = "0.7", features = ["codec"]}
 
-io-lifetimes = {version = "0.7"}
+io-lifetimes = {version = "1.0.5"}
 tracing = {version = "0.1", default-features = false, features = ["std"]}
 tracing-subscriber = {version = "0.3", default-features = false, features = ["std", "fmt"]}
-uuid = {version = "0.8.2", features = ["v4"]}
+uuid = {version = "1.3", features = ["v4"]}
 
 [features]
 default = []
